requirements.txt
streamlit>=1.36
python-docx>=1.1.0
requests>=2.31
numpy>=1.26
scikit-learn>=1.3
